LPS-108030 Move to internal packages named based on implemented interfaces
