+++ modules/saas/apps/saml-saas/src/main/java/com/liferay/saml/saas/internal/jaxrs/application/ImportSamlSaasApplication.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.saml.saas.application;
+package com.liferay.saml.saas.internal.jaxrs.application;
 
 import com.liferay.expando.kernel.model.ExpandoBridge;
 import com.liferay.portal.kernel.exception.PortalException;
@@ -35,8 +35,8 @@ import com.liferay.saml.persistence.service.SamlSpIdpConnectionLocalService;
 import com.liferay.saml.runtime.configuration.SamlConfiguration;
 import com.liferay.saml.runtime.configuration.SamlProviderConfigurationHelper;
 import com.liferay.saml.runtime.credential.KeyStoreManager;
-import com.liferay.saml.saas.constants.JSONKeys;
-import com.liferay.saml.saas.util.SymmetricEncryptor;
+import com.liferay.saml.saas.internal.constants.JSONKeys;
+import com.liferay.saml.saas.internal.util.SymmetricEncryptor;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
